+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Fix Protontricks crash when Proton installation is incomplete
  - Fix Protontricks crash when both Flatpak and non-Flatpak versions of Steam are installed
  - Fix duplicate log messages when using `protontricks-launch`
+ - Fix error dialog not being displayed when using `protontricks-launch`
 
 ## [1.7.0] - 2022-01-08
 ### Changed